Update EE6FeatureReplacementAction to have an isEnabled method
- Add isEnabled() method override to test if running repeat with an OL
install vs an WL install.  If OL is installed, it should not be enabled.
This change standardizes this checking for everyone and everyone doesn't
need to do their own override.
- Update subclasses of EE6FeatureReplacementAction to not need
isEnabled() method and for servlet scenario just remove the subclass.
- Add EE6_FEATURES() method to FeatureReplacementAction class just like
we have for the EE levels.  Also update method javadoc comments to
remove the old information and make it similar for all EE levels.
