+++ dev/com.ibm.ws.jpa.tests.spec10.embeddable_jpa_2.0_fat/fat/src/com/ibm/ws/jpa/tests/embeddable/RepeatWithJPA20.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2023 IBM Corporation and others.
+ * Copyright (c) 2023, 2025 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -13,27 +13,12 @@
 
 package com.ibm.ws.jpa.tests.embeddable;
 
-import java.io.File;
-
 import com.ibm.ws.testtooling.jpaprovider.JPAPersistenceProvider;
 
-import componenttest.common.apiservices.Bootstrap;
 import componenttest.rules.repeater.EE6FeatureReplacementAction;
 
 public class RepeatWithJPA20 extends EE6FeatureReplacementAction {
     public static final String ID = "JPA20";
-
-    @Override
-    public boolean isEnabled() {
-        try {
-            Bootstrap b = Bootstrap.getInstance();
-            String installRoot = b.getValue("libertyInstallPath");
-            File jpa20Feature = new File(installRoot + "/lib/features/com.ibm.websphere.appserver.jpa-2.0.mf");
-            return jpa20Feature.exists() && super.isEnabled();
-        } catch (Exception e) {
-            return false;
-        }
-    }
 
     @Override
     public String toString() {
