+++ dev/com.ibm.ws.jpa.tests.spec20_jpa_2.0_fat/fat/src/com/ibm/ws/jpa/tests/spec20/RepeatWithJPA20.java
@@ -1,10 +1,10 @@
 /*******************************************************************************
- * Copyright (c) 2022 IBM Corporation and others.
+ * Copyright (c) 2022, 2025 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
@@ -13,11 +13,8 @@
 
 package com.ibm.ws.jpa.tests.spec20;
 
-import java.io.File;
-
 import com.ibm.ws.testtooling.jpaprovider.JPAPersistenceProvider;
 
-import componenttest.common.apiservices.Bootstrap;
 import componenttest.custom.junit.runner.Mode.TestMode;
 import componenttest.rules.repeater.EE6FeatureReplacementAction;
 
@@ -46,17 +43,5 @@ public class RepeatWithJPA20 extends EE6FeatureReplacementAction {
     @Override
     public String getID() {
         return ID;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        try {
-            Bootstrap b = Bootstrap.getInstance();
-            String installRoot = b.getValue("libertyInstallPath");
-            File jpa20Feature = new File(installRoot + "/lib/features/com.ibm.websphere.appserver.jpa-2.0.mf");
-            return jpa20Feature.exists() && super.isEnabled();
-        } catch (Exception e) {
-            return false;
-        }
     }
 }
